# SWAGGER-NVSL-APIS.yaml
openapi: 3.0.3
info:
  title: NVSL API
  version: 1.0.0
  description: |
    Documentação consolidada das APIs do NVSL.
    Base URL padrão: `http://localhost:8081/api`.

    Rotas **web** na raiz da aplicação (`http://localhost:8081`, cadastro MGI):
    - `GET /` — status da API
    - `GET /redirect-gov` — callback OAuth2/OIDC (redirect_uri)
    - `GET /logout` — redireciona ao login do frontend (`?from=logout`); revogação do token: `POST /api/auth/logout`
servers:
  - url: http://localhost:8081/api
    description: Ambiente local (Docker)
...
paths:
...
  /gerenciar-perfis:
    get:
      tags: [Gerenciar Perfis]
      summary: Lista perfis para gestão
      security:
        - bearerAuth: []
      parameters:
        - name: nome
          in: query
          schema: { type: string }
        - name: esfera
          in: query
          schema:
            type: string
            enum: [federal, estadual, municipal]
        - name: status
          in: query
          schema:
            type: string
            enum: [ativo, inativo]
      responses:
        "200":
          description: Lista de perfis
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PerfilGerenciar"
        "401":
          description: Não autenticado
        "403":
          description: Acesso não permitido

    post:
      tags: [Gerenciar Perfis]
      summary: Cadastra novo perfil
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GerenciarPerfilStoreRequest"
      responses:
        "201":
          description: Perfil cadastrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    $ref: "#/components/schemas/PerfilGerenciar"
        "401":
          description: Não autenticado
        "403":
          description: Acesso não permitido
        "422":
          description: Erro de validação
  /gerenciar-perfis/permissoes:
    get:
      tags: [Gerenciar Perfis]
      summary: Lista permissões disponíveis para o usuário
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de permissões
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PermissaoGerenciar"
        "401":
          description: Não autenticado
  /gerenciar-perfis/hierarquia:
    get:
      tags: [Gerenciar Perfis]
      summary: Retorna hierarquia de atuação do usuário
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Esferas permitidas para gestão de perfis
          content:
            application/json:
              schema:
                type: object
                properties:
                  esfera_usuario:
                    type: string
                    enum: [federal, estadual, municipal]
                  esferas_permitidas:
                    type: array
                    items:
                      type: string
                      enum: [federal, estadual, municipal]
        "401":
          description: Não autenticado
  /gerenciar-perfis/{id}:
    get:
      tags: [Gerenciar Perfis]
      summary: Detalha perfil por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Perfil detalhado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PerfilGerenciar"
        "401":
          description: Não autenticado
        "403":
          description: Acesso não permitido
        "404":
          description: Perfil não encontrado

    put:
      tags: [Gerenciar Perfis]
      summary: Atualiza perfil existente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GerenciarPerfilStoreRequest"
      responses:
        "200":
          description: Perfil atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    $ref: "#/components/schemas/PerfilGerenciar"
        "401":
          description: Não autenticado
        "403":
          description: Acesso não permitido
        "404":
          description: Perfil não encontrado
        "422":
          description: Erro de validação
  /gerenciar-perfis/{id}/historico:
    get:
      tags: [Gerenciar Perfis]
      summary: Lista histórico de alterações do perfil
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Histórico do perfil
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/HistoricoPerfilItem"
        "401":
          description: Não autenticado
        "403":
          description: Acesso não permitido
        "404":
          description: Perfil não encontrado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
...
    PermissaoGerenciar:
      type: object
      properties:
        id: { type: integer }
        funcionalidade: { type: string }
        nome_acao: { type: string }
        descricao: { type: string, nullable: true }
    PerfilGerenciar:
      type: object
      properties:
        id: { type: integer }
        nome: { type: string }
        descricao: { type: string, nullable: true }
        esfera:
          type: string
          enum: [federal, estadual, municipal]
        status:
          type: string
          enum: [ativo, inativo]
        permissoes:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              funcionalidade: { type: string }
              nome_acao: { type: string }
        created_at: { type: string, format: date-time, nullable: true }
    GerenciarPerfilStoreRequest:
      type: object
      required: [nome, esfera, status]
      properties:
        nome: { type: string, maxLength: 100 }
        descricao: { type: string, nullable: true, maxLength: 255 }
        esfera:
          type: string
          enum: [federal, estadual, municipal]
        status:
          type: string
          enum: [ativo, inativo]
        permissoes:
          type: array
          items:
            type: integer
    HistoricoPerfilItem:
      type: object
      properties:
        id: { type: integer }
        data_hora: { type: string, example: "25/03/2026 11:14:32" }
        usuario: { type: string }
        perfil: { type: string }
        atualizacao: { type: string }
        action: { type: string }
...